# build(optimizer): fix stack too deeeeeeep