## Before

```
╭----------------------+------------------+-------------------+--------------------+---------------------╮
| Contract             | Runtime Size (B) | Initcode Size (B) | Runtime Margin (B) | Initcode Margin (B) |
+========================================================================================================+
| Address              | 85               | 160               | 24,491             | 48,992              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| AtmAuction           | 9,402            | 10,243            | 15,174             | 38,909              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| BondAuction          | 11,708           | 12,563            | 12,868             | 36,589              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Checkpoints          | 85               | 160               | 24,491             | 48,992              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Deposit              | 4,459            | 5,407             | 20,117             | 43,745              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| DoubleEndedQueue     | 85               | 160               | 24,491             | 48,992              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| DutchAuction         | 9,000            | 9,823             | 15,576             | 39,329              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| ECDSA                | 85               | 160               | 24,491             | 48,992              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Errors               | 85               | 160               | 24,491             | 48,992              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| EthStrategy          | 49,352           | 58,937            | -24,776            | -9,785              |
```

## After

```
╭----------------------+------------------+-------------------+--------------------+---------------------╮
| Contract             | Runtime Size (B) | Initcode Size (B) | Runtime Margin (B) | Initcode Margin (B) |
+========================================================================================================+
| Address              | 8                | 39                | 24,568             | 49,113              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| AtmAuction           | 6,286            | 6,873             | 18,290             | 42,279              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| BondAuction          | 7,573            | 8,174             | 17,003             | 40,978              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Checkpoints          | 8                | 39                | 24,568             | 49,113              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Deposit              | 3,821            | 4,529             | 20,755             | 44,623              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| DoubleEndedQueue     | 8                | 39                | 24,568             | 49,113              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| DutchAuction         | 5,797            | 6,360             | 18,779             | 42,792              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| ECDSA                | 8                | 39                | 24,568             | 49,113              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| Errors               | 8                | 39                | 24,568             | 49,113              |
|----------------------+------------------+-------------------+--------------------+---------------------|
| EthStrategy          | 34,230           | 41,160            | -9,654             | 7,992               |
|----------------------+------------------+-------------------+--------------------+---------------------|

```