@sambacha 2025-02-14
-9,654
left

we have the technology