@dangerousfood 2025-02-27
Fixed in fix/eip-170 branch